Added arguments for selecting dataset and network as described in the assignment #4 #5
See get_args() function or type python main.py --help. The values you can specify for the new arguments (--dataset, --network) are the same as the "type" attribute in the yml file. That way, we can specify many datasets and network structures in the yml and select which one to use by providing the appropriate argument.